Add 'expnum' to filenames for 'stack' and related products.
The Butler automatically looks up all information necessary
to construct the full POSIX filename, even if only given a partial ID.
But it doesn't lo9ok up anything else.  Putting the expnum in the
filename for the stack images and related products means that you
don't have to specify the expnum when doing something with one of those files.

While not necessary to create uniqe _filenames_ where we use 'field' and 'filter',
the expnum is nused for constructing unique exposure IDs,
which has to be an integer.